backend/requirements.txt
fastapi
uvicorn[standard]
python-dateutil
pytest
httpx
